+++ packages/apollo-server-core/src/runQuery.ts
@@ -135,6 +135,7 @@ function doRunQuery(options: QueryOptions): Promise<GraphQLResponse> {
     variables: options.variables,
     persistedQueryHit: options.persistedQueryHit,
     persistedQueryRegister: options.persistedQueryRegister,
+    context,
   });
   return Promise.resolve()
     .then(
@@ -278,7 +279,10 @@ function doRunQuery(options: QueryOptions): Promise<GraphQLResponse> {
       throw err;
     })
     .then((graphqlResponse: GraphQLResponse) => {
-      const response = extensionStack.willSendResponse({ graphqlResponse });
+      const response = extensionStack.willSendResponse({
+        graphqlResponse,
+        context,
+      });
       requestDidEnd();
       return response.graphqlResponse;
     });