+++ packages/graphql-extensions/src/index.ts
@@ -40,6 +40,7 @@ export class GraphQLExtension<TContext = any> {
     variables?: { [key: string]: any };
     persistedQueryHit?: boolean;
     persistedQueryRegister?: boolean;
+    context: TContext;
   }): EndHandler | void;
   public parsingDidStart?(o: { queryString: string }): EndHandler | void;
   public validationDidStart?(): EndHandler | void;
@@ -49,7 +50,8 @@ export class GraphQLExtension<TContext = any> {
 
   public willSendResponse?(o: {
     graphqlResponse: GraphQLResponse;
-  }): void | { graphqlResponse: GraphQLResponse };
+    context: TContext;
+  }): void | { graphqlResponse: GraphQLResponse; context: TContext };
 
   public willResolveField?(
     source: any,
@@ -78,6 +80,7 @@ export class GraphQLExtensionStack<TContext = any> {
     variables?: { [key: string]: any };
     persistedQueryHit?: boolean;
     persistedQueryRegister?: boolean;
+    context: TContext;
   }): EndHandler {
     return this.handleDidStart(
       ext => ext.requestDidStart && ext.requestDidStart(o),
@@ -104,7 +107,8 @@ export class GraphQLExtensionStack<TContext = any> {
 
   public willSendResponse(o: {
     graphqlResponse: GraphQLResponse;
-  }): { graphqlResponse: GraphQLResponse } {
+    context: TContext;
+  }): { graphqlResponse: GraphQLResponse; context: TContext } {
     let reference = o;
     // Reverse the array, since this is functions as an end handler
     [...this.extensions].reverse().forEach(extension => {